add initial_page_data test
