+++ corehq/apps/hqwebapp/templatetags/tests/rendered/initial_page_data.html
@@ -1,0 +1,3 @@
+<body>
+<div data-name="scalar" data-value="estrella"></div><div data-name="special_chars" data-value="here&#39;s &quot;something&quot; irritating to parse &amp; deal with"></div><div data-name="list" data-value="[1, 2, 3]"></div><div data-name="maps" data-value="[{&quot;hen&quot;: &quot;brood&quot;}, {&quot;nightingale&quot;: &quot;watch&quot;}, {&quot;quail&quot;: &quot;bevy&quot;}, {&quot;starling&quot;: &quot;murmuration&quot;}]"></div>
+</body>
