+++ corehq/apps/hqwebapp/templatetags/tests/templates/initial_page_data.html
@@ -1,0 +1,11 @@
+{% load hq_shared_tags %}
+
+{% initial_page_data 'scalar' scalar %}
+{% initial_page_data 'special_chars' special_chars %}
+{% initial_page_data 'list' list %}
+{% initial_page_data 'maps' maps %}
+
+<body>
+{% block initial_page_data %}
+{% endblock %}
+</body>
